取关调试

+++ BlogBackEnd/src/main/java/com/cqucs/blogbackend/controller/UserController.java
@@ -212,7 +212,12 @@ public class UserController {
     @DeleteMapping("/unfollow/{use_u_id}")
     public OperateResult unfollow(@PathVariable Integer use_u_id,Integer u_id){
         try{
-            String sql = "delete from follow where u_id=? and use_u_id=?";
+            String sql = "select count(*) from follow where u_id=? and use_u_id=?";
+            Integer count = jdbcTemplate.queryForObject(sql,Integer.class,u_id,use_u_id);
+            if(count==0){
+                return new OperateResult(500,"关系不存在",null) ;
+            }
+            sql = "delete from follow where u_id=? and use_u_id=?";
             Object[] args = {u_id,use_u_id};
             jdbcTemplate.update(sql,args);
             return new OperateResult(200,"取消关注成功",null) ;
